[Step 3][fcntl.h] Types and structures.
