+++ src/unix/notbsd/linux/common/b64/mod.rs
@@ -2,6 +2,23 @@
 pub type c_long = i64;
 pub type c_ulong = u64;
 
+/* Header <fcntl.h> */
+cfg_if! {
+    if #[cfg(feature = "file_offset64")] {
+        type flock = ::flock64;
+    } else {
+        s! {
+            pub struct flock {
+                pub l_type: ::c_short,
+                pub l_whence: ::c_short,
+                pub l_start: ::off_t,
+                pub l_len: ::off_t,
+                pub l_pid: ::pid_t,
+            }
+        }
+    }
+}
+
 cfg_if! {
     if #[cfg(target_arch = "aarch64")] {
         mod aarch64;
