+++ src/unix/notbsd/linux/mips.rs
@@ -2,7 +2,6 @@ pub type clock_t = i32;
 pub type time_t = i32;
 pub type suseconds_t = i32;
 pub type wchar_t = i32;
-pub type off_t = i32;
 pub type ino_t = u32;
 pub type blkcnt_t = i32;
 pub type blksize_t = i32;
@@ -158,16 +157,6 @@ s! {
         pub c_lflag: ::tcflag_t,
         pub c_line: ::cc_t,
         pub c_cc: [::cc_t; ::NCCS],
-    }
-
-    pub struct flock {
-        pub l_type: ::c_short,
-        pub l_whence: ::c_short,
-        pub l_start: ::off_t,
-        pub l_len: ::off_t,
-        pub l_sysid: ::c_long,
-        pub l_pid: ::pid_t,
-        pad: [::c_long; 4],
     }
 
     pub struct sysinfo {
