+++ src/unix/notbsd/linux/other/b32/mod.rs
@@ -4,7 +4,6 @@ pub type clock_t = i32;
 pub type time_t = i32;
 pub type suseconds_t = i32;
 pub type ino_t = u32;
-pub type off_t = i32;
 pub type blkcnt_t = i32;
 pub type __fsword_t = i32;
