+++ src/unix/notbsd/linux/other/b64/mod.rs
@@ -4,7 +4,6 @@ pub type clock_t = i64;
 pub type time_t = i64;
 pub type suseconds_t = i64;
 pub type ino_t = u64;
-pub type off_t = i64;
 pub type blkcnt_t = i64;
 pub type __fsword_t = ::c_long;
